+++ pkg/server/singleprocess/service_logs_test.go
@@ -379,11 +379,12 @@ func TestServiceGetLogStream_depPlugin(t *testing.T) {
 	// Should have the job to spawn the logs plugin
 	var job *pb.Job
 	require.Eventually(t, func() bool {
-		jobs, err := testServiceImpl(impl).state(ctx).JobList(&pb.ListJobsRequest{})
+		jobsResp, err := client.ListJobs(ctx, &pb.ListJobsRequest{})
 		if err != nil {
 			t.Logf("error getting job list: %s", err)
 			return false
 		}
+		jobs := jobsResp.Jobs
 
 		if len(jobs) == 1 {
 			job = jobs[0]
